+++ packages/plugin-gallery/web/src/components/gallery-controls/radio-groups.js
@@ -19,7 +19,7 @@ export const ThumbnailResize = props => {
   const thumbnailResizeLabel = t('GallerySettings_Radios_Thumbnail_Resize');
   const cropLabel = t('GallerySettings_Radios_Crop');
   const fitLabel = t('GallerySettings_Radios_Fit');
-  const RadioGroupComponent = SelectRadioGroupComponent(experiments?.newSettingsUi?.enabled);
+  const RadioGroupComponent = SelectRadioGroupComponent(experiments?.newSettingsModals?.enabled);
 
   return (
     <RadioGroupComponent
@@ -41,7 +41,7 @@ export const TitleButtonPlacement = props => {
   const titleButtonPlacementLabel = t('GallerySettings_Radios_Title_Button_Placement');
   const underneathLabel = t('GallerySettings_Radios_Underneath');
   const onHoverLabel = t('GallerySettings_Radios_On_Hover');
-  const RadioGroupComponent = SelectRadioGroupComponent(experiments?.newSettingsUi?.enabled);
+  const RadioGroupComponent = SelectRadioGroupComponent(experiments?.newSettingsModals?.enabled);
 
   return (
     <RadioGroupComponent
@@ -61,7 +61,7 @@ export const ImageOrientation = props => {
   const imageOrientationLabel = t('GallerySettings_Radios_Image_Orientation');
   const verticalLabel = t('GallerySettings_Radios_Vertical');
   const horizontalLabel = t('GallerySettings_Radios_Horizontal');
-  const RadioGroupComponent = SelectRadioGroupComponent(experiments?.newSettingsUi?.enabled);
+  const RadioGroupComponent = SelectRadioGroupComponent(experiments?.newSettingsModals?.enabled);
 
   return (
     <RadioGroupComponent
@@ -87,7 +87,7 @@ export const ScrollDirection = props => {
   const scrollDirectionLabel = t('GallerySettings_Radios_Scroll_Direction');
   const verticalLabel = t('GallerySettings_Radios_Vertical');
   const horizontalLabel = t('GallerySettings_Radios_Horizontal');
-  const RadioGroupComponent = SelectRadioGroupComponent(experiments?.newSettingsUi?.enabled);
+  const RadioGroupComponent = SelectRadioGroupComponent(experiments?.newSettingsModals?.enabled);
 
   return (
     <RadioGroupComponent