+++ packages/plugin-social-polls/web/src/components/settings/poll-settings-section/PollSettingsSection.jsx
@@ -18,7 +18,7 @@ export class PollSettingsSection extends Component {
 
   modalsWithEditorCommands = this.props.experiments.modalBaseActionHoc?.enabled;
 
-  useNewSettingsUi = !!this.props.experiments?.newSettingsUi?.enabled;
+  useNewSettingsUi = !!this.props.experiments?.newSettingsModals?.enabled;
 
   VOTE_ROLE_OPTIONS = [
     {